#25: create road sign component
