+++ app/src/assets/css/style.css
@@ -235,4 +235,36 @@ body {
   height: 5rem;
   border-radius: 0.5rem;
   border: 1px solid white;
+}
+.c__road-event-init {
+  display: flex;
+  width: 30rem;
+  height: 5rem;
+  padding: 0.5rem;
+  position: absolute;
+  top: 5%;
+  justify-content: center;
+  align-items: center;
+  left: calc((100vw - 30rem) / 2);
+  background-color: rgb(252, 255, 45);
+  border-radius: 0.2rem;
+  z-index: 5;
+}
+.c__road-event-init .event-text {
+  display: inline;
+  margin-right: 0.5rem;
+  width: 30%;
+  text-align: center;
+}
+.c__road-event-init .event-img {
+  display: inline-block;
+  width: 40%;
+  height: 4rem;
+}
+.c__road-event-init .event-img img {
+  width: 100%;
+  height: 100%;
+  -o-object-fit: contain;
+     object-fit: contain;
+  overflow: hidden;
 }/*# sourceMappingURL=style.css.map */
