+++ app/src/assets/scss/components/_components.scss
@@ -104,5 +104,37 @@
         }
 
     }
+    &__road-event-init{
+        display: flex;
+        width: 30rem;
+        height: 5rem;
+        padding: 0.5rem;
+        position: absolute;
+        top: 5%;
+        justify-content: center;
+        align-items: center;
+        left: calc((100vw - 30rem) / 2);
+        background-color: rgb(252, 255, 45);
+        border-radius: 0.2rem;
+        z-index: 5;
+
+        .event-text {
+            display: inline;
+            margin-right: 0.5rem;
+            width: 30%;
+            text-align: center;
+        }
+        .event-img {
+            display: inline-block;
+            width: 40%;
+            height: 4rem;
+            img {
+                width: 100%;
+                height: 100%;
+                object-fit: contain;
+                overflow: hidden;
+            }
+        }
+    }
 }
