--- src/assets/css/index.css
@@ -1,141 +1,0 @@
-.u__container--medium {
-  margin: 10rem auto;
-  background-color: #D5E8D4;
-  width: 50rem;
-  border-radius: 0.5rem;
-  padding: 2rem;
-}
-.u__layout--text--center {
-  text-align: center;
-}
-.u__layout--block--center {
-  justify-content: center;
-}
-.u__button--large {
-  display: block;
-  margin: 0.5rem auto;
-  width: 20rem;
-  height: 2.5rem;
-  line-height: 2.5rem;
-  border-radius: 0.7rem;
-  border: none;
-  color: white;
-  background-color: #60A917;
-  cursor: pointer;
-}
-.u__button--large:hover {
-  background-color: #76cd1f;
-}
-.u__p--center {
-  margin: 0.7rem auto;
-  text-align: center;
-  font-weight: 700;
-}
-.u__margin--all--md {
-  padding: 2rem;
-}
-.u__margin--bottom--md {
-  margin-bottom: 2rem;
-}
-.u__margin--bottom--sm {
-  margin-bottom: 1rem;
-}
-.u__visibility--hide {
-  display: none !important;
-}/*# sourceMappingURL=index.css.map */font-size: 100%;
-}
-
-*, *::before, *::after {
-  box-sizing: inherit;
-}
-
-body {
-  background-color: black;
-  margin: 0;
-  padding: 0;
-  font-family: "Rubik", sans-serif;
-}/*# sourceMappingURL=index.css.map */00;
-}
-.u__margin--all--md {
-  padding: 2rem;
-}
-.u__margin--bottom--md {
-  margin-bottom: 2rem;
-}
-.u__margin--bottom--sm {
-  margin-bottom: 1rem;
-}
-.u__visibility--hide {
-  display: none !important;
-}
-
-html {
-  box-sizing: border-box;
-  font-size: 100%;
-}
-
-*, *::before, *::after {
-  box-sizing: inherit;
-}
-
-body {
-  background-color: black;
-  margin: 0;
-  padding: 0;
-  font-family: "Rubik", sans-serif;
-}
-
-.p__modeSelector--container {
-  background-color: black;
-  color: rgb(146, 255, 146);
-  display: flex;
-  justify-content: center;
-}
-.p__modeSelector--inner-card {
-  margin: 5rem auto;
-  padding: 2rem;
-  text-align: center;
-  border-radius: 1rem;
-  background-color: white;
-  color: black;
-  width: 40rem;
-}
-.p__modeSelector--form button {
-  width: 10rem;
-  height: 4rem;
-  line-height: 4rem;
-  margin: 0.5rem;
-  color: white;
-  background-color: green;
-  border: none;
-  border-radius: 1rem;
-  cursor: pointer;
-}
-.p__modeSelector--form button:hover {
-  background-color: rgb(4, 179, 4);
-}
-.p__voice-mode__image-container {
-  margin: auto;
-  width: 10rem;
-  height: 8rem;
-  cursor: pointer;
-}
-.p__voice-mode__recording-img {
-  width: 10rem;
-  height: 8rem;
-  border-radius: 1rem;
-}
-
-.but {
-  box-sizing: border-box;
-  border: 1px solid white;
-}
-
-.span {
-  padding: 1rem;
-}
-
-.directions {
-  display: block;
-  margin-top: 1rem;
-}/*# sourceMappingURL=index.css.map */
