--- src/assets/css/style.css
@@ -1,238 +1,0 @@
-.u__container--medium {
-  margin: 10rem auto;
-  background-color: #D5E8D4;
-  width: 50rem;
-  border-radius: 0.5rem;
-  padding: 2rem;
-}
-.u__layout--text--center {
-  text-align: center;
-}
-.u__layout--block--center {
-  justify-content: center;
-}
-.u__button--large {
-  display: block;
-  margin: 0.5rem auto;
-  width: 20rem;
-  height: 2.5rem;
-  line-height: 2.5rem;
-  border-radius: 0.7rem;
-  border: none;
-  color: white;
-  background-color: #60A917;
-  cursor: pointer;
-}
-.u__button--large:hover {
-  background-color: #76cd1f;
-}
-.u__p--center {
-  margin: 0.7rem auto;
-  text-align: center;
-  font-weight: 700;
-}
-.u__margin--all--md {
-  padding: 2rem;
-}
-.u__margin--bottom--md {
-  margin-bottom: 2rem;
-}
-.u__margin--bottom--sm {
-  margin-bottom: 1rem;
-}
-.u__visibility--hide {
-  display: none !important;
-}
-
-html {
-  box-sizing: border-box;
-  font-size: 100%;
-}
-
-*, *::before, *::after {
-  box-sizing: inherit;
-}
-
-body {
-  background-color: black;
-  margin: 0;
-  padding: 0;
-  font-family: "Rubik", sans-serif;
-}
-
-.pg__hs--wrapper {
-  display: flex;
-  justify-content: center;
-  padding-top: 40vh;
-  height: 100vh;
-  background-color: black;
-  /* box-sizing: border-box; */
-  color: white;
-}
-.pg__hs .inner-div {
-  display: inline-block;
-  /* height: 200px; */
-}
-.pg__hs div.mic2 {
-  padding: 2rem;
-}
-.pg__hs .mic {
-  height: 50px;
-  display: inline-block;
-  border: 0.1px solid white;
-  border-radius: 2rem;
-}
-.pg__hs .welcome-text {
-  display: block;
-}
-
-.p__modeSelector--container {
-  background-color: black;
-  color: rgb(146, 255, 146);
-  display: flex;
-  justify-content: center;
-}
-.p__modeSelector--inner-card {
-  margin: 5rem auto;
-  padding: 2rem;
-  text-align: center;
-  border-radius: 1rem;
-  background-color: white;
-  color: black;
-  width: 40rem;
-}
-.p__modeSelector--form button {
-  width: 10rem;
-  height: 4rem;
-  line-height: 4rem;
-  margin: 0.5rem;
-  color: white;
-  background-color: green;
-  border: none;
-  border-radius: 1rem;
-  cursor: pointer;
-}
-.p__modeSelector--form button:hover {
-  background-color: rgb(4, 179, 4);
-}
-.p__voice-mode__image-container {
-  margin: auto;
-  width: 10rem;
-  height: 8rem;
-  cursor: pointer;
-}
-.p__voice-mode__recording-img {
-  width: 10rem;
-  height: 8rem;
-  border-radius: 1rem;
-}
-
-.but {
-  box-sizing: border-box;
-  border: 1px solid white;
-}
-
-.span {
-  padding: 1rem;
-}
-
-.directions {
-  display: block;
-  margin-top: 1rem;
-}
-
-.c__text-mode--input {
-  width: 20rem;
-  height: 5rem;
-  padding: 1rem;
-  border: none;
-}
-.c__search-result {
-  display: flex;
-  height: 100vh;
-  width: 100vw;
-}
-.c__search-result__section {
-  width: 75%;
-  height: 100%;
-}
-.c__search-result__section .dest-img {
-  position: absolute;
-  bottom: 3rem;
-  right: 28rem;
-  width: 20rem;
-  height: 10rem;
-}
-.c__search-result__section .dest-img img {
-  max-width: 100%;
-  max-height: 100%;
-  border-radius: 1rem;
-}
-.c__search-result__aside {
-  width: 25%;
-  height: 100%;
-  padding: 1rem;
-  background-color: white;
-}
-.c__navigator {
-  display: flex;
-  flex-direction: column;
-  position: absolute;
-  top: 40%;
-  left: 0;
-  background-color: rgb(213, 232, 212);
-  border-radius: 0.2rem;
-  z-index: 999;
-}
-.c__navigator .route-span {
-  width: 2rem;
-  margin: 0.2rem;
-  padding-left: 0.2rem;
-  background-color: rgba(213, 232, 212, 0.7);
-  border-radius: 0.2rem;
-  border: 0.1px solid green;
-}
-.c__navigator .route-span:hover {
-  background-color: rgb(213, 232, 212);
-  cursor: pointer;
-  color: white;
-  background: green;
-}
-.c__road-event {
-  display: flex;
-  width: 30rem;
-  height: 5rem;
-  padding: 0.5rem;
-  position: absolute;
-  top: 5%;
-  justify-content: center;
-  align-items: center;
-  left: calc(30% - 4rem);
-  background-color: rgb(252, 255, 45);
-  border-radius: 0.2rem;
-  z-index: 5;
-}
-.c__road-event .event-text {
-  display: inline;
-  margin-right: 0.5rem;
-  width: 30%;
-  text-align: center;
-}
-.c__road-event .event-img {
-  display: inline-block;
-  width: 40%;
-  height: 4rem;
-}
-.c__road-event .event-img img {
-  width: 100%;
-  height: 100%;
-  -o-object-fit: contain;
-     object-fit: contain;
-  overflow: hidden;
-}
-.c__road-event--img-container {
-  width: 5rem;
-  height: 5rem;
-  border-radius: 0.5rem;
-  border: 1px solid white;
-}/*# sourceMappingURL=style.css.map */
